// frontend/src/app/swagger/model/layoutReadDto.ts
/**
 * S2Retro.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ColumnReadDto } from './columnReadDto';


export interface LayoutReadDto { 
    id?: string;
    name?: string | null;
    isArchived?: boolean;
    columns?: Array<ColumnReadDto> | null;
}