// frontend/src/app/swagger/model/layoutCreateDto.ts
/**
 * S2Retro.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ColumnCreateDto } from './columnCreateDto';


export interface LayoutCreateDto { 
    name?: string | null;
    columns?: Array<ColumnCreateDto> | null;
}